refactor: make `earthdawn` utilities `helpers`
