+++ module/data/other/character-generation.mjs
@@ -4,7 +4,7 @@ import SparseDataModel from "../abstract/sparse-data-model.mjs";
 import { SYSTEM_TYPES } from "../../constants/constants.mjs";
 import * as ACTORS from "../../config/actors.mjs";
 import * as LEGEND from "../../config/legend.mjs";
-import { getAttributeStepFromValue } from "../../utils/earthdawn.mjs";
+import { getAttributeStepFromValue } from "../../helpers/earthdawn.mjs";
 import { sum } from "../../utils/math.mjs";
 import { getSingleGlobalItemByEdid, prefixKeysForDeletion } from "../../helpers/document.mjs";
 import { filterObject, mapObject } from "../../utils/object.mjs";
