+++ module/utils/_module.mjs
@@ -1,9 +1,9 @@
-export * as earthdawn from "./earthdawn.mjs";
+export * as earthdawn from "../helpers/earthdawn.mjs";
 export * as math from "./math.mjs";
 export * as object from "./object.mjs";
 export * as string from "./string.mjs";
 
-export * from "./earthdawn.mjs";
+export * from "../helpers/earthdawn.mjs";
 export * from "./math.mjs";
 export * from "./object.mjs";
 export * from "./string.mjs";
